requirements.txt
asttokens==2.0.5
backcall==0.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
decorator==5.1.1
executing==0.8.3
google-api-core==2.4.0
google-api-python-client==2.35.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
httplib2==0.20.2
idna==3.3
ipython==8.4.0
jedi==0.18.1
matplotlib-inline==0.1.3
numpy==1.22.1
oauthlib==3.1.1
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
protobuf==3.19.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
stack-data==0.3.0
tabulate==0.8.9
traitlets==5.3.0
uritemplate==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
